requirements.txt
matplotlib==3.8.0
mplfinance==0.12.10b0
Pillow==10.0.0
Pillow==10.2.0
streamlit==1.29.0
ultralytics==8.1.6
yfinance==0.2.36